Factored out express app configuration...
... from running the server.

The `server.ts` is just there as a basic server implementation, but
`metadata-transformer` is intended to be used as a library and the new
`src/index.ts` entrypoint now reflects that.

This change required me to modify the transformer implementations to get
a web3 client injected into them. I made all exports non-default so that
users have to explicitly import the functions which *produce*
transformers - this will make it less likely for them to try to use the
`createXTransformer` function in place of an `XTransformer`.

Resolves https://github.com/moonstream-to/metadata-transformer/issues/2, https://github.com/moonstream-to/metadata-transformer/issues/3
